refactor(api-service): streamline OAuth URL generation and remove unused code
- Updated OAuth URL generation in multiple use cases to utilize the API_ROOT_URL environment variable instead of hardcoded URLs.
- Removed redundant ZIP building logic from the AzureSetupOauthCallback and teams-app-package components, leveraging a shared utility function instead.
- Cleaned up commented-out code and improved overall readability of the affected files.
