+++ apps/api/src/app/integrations/usecases/generate-chat-oath-url/generate-msteams-oath-url/generate-msteams-oauth-url.usecase.ts
@@ -194,7 +194,7 @@ export class GenerateMsTeamsOauthUrl {
 
     const baseUrl = process.env.API_ROOT_URL.replace(/\/$/, ''); // Remove trailing slash
 
-    return `https://gosha.ngrok.app${CHAT_OAUTH_CALLBACK_PATH}`;
+    return `${baseUrl}${CHAT_OAUTH_CALLBACK_PATH}`;
     // return `${baseUrl}${CHAT_OAUTH_CALLBACK_PATH}`;
   }
