+++ apps/api/src/app/integrations/usecases/generate-msteams-arm-template/generate-msteams-arm-template.usecase.ts
@@ -80,8 +80,7 @@ export class GenerateMsTeamsArmTemplate {
   }
 
   private buildTemplateApiUrl(integrationId: string, sig: string, exp: number): string {
-    // const base = (process.env.API_ROOT_URL ?? '').replace(/\/$/, '');
-    const base = 'https://gosha.ngrok.app';
+    const base = (process.env.API_ROOT_URL ?? '').replace(/\/$/, '');
 
     return `${base}/v1/integrations/${integrationId}/msteams-arm-template?sig=${sig}&exp=${exp}`;
   }
